+++ src/access/domain/repositories/IAccessRepository.ts
@@ -1,6 +1,14 @@
 import { GuestbookResponseDTO } from '../dtos/GuestbookResponseDTO'
 
 export interface IAccessRepository {
+  getSignedDatafileDownloadUrl(fileId: number | string): Promise<string>
+
+  getSignedDatafilesDownloadUrl(fileIds: string | Array<number | string>): Promise<string>
+
+  getSignedDatasetDownloadUrl(datasetId: number | string): Promise<string>
+
+  getSignedDatasetVersionDownloadUrl(datasetId: number | string, versionId: string): Promise<string>
+
   submitGuestbookForDatafileDownload(
     fileId: number | string,
     guestbookResponse: GuestbookResponseDTO