+++ src/access/index.ts
@@ -1,4 +1,8 @@
 import { AccessRepository } from './infra/repositories/AccessRepository'
+import { GetSignedDatafileDownloadUrl } from './domain/useCases/GetSignedDatafileDownloadUrl'
+import { GetSignedDatafilesDownloadUrl } from './domain/useCases/GetSignedDatafilesDownloadUrl'
+import { GetSignedDatasetDownloadUrl } from './domain/useCases/GetSignedDatasetDownloadUrl'
+import { GetSignedDatasetVersionDownloadUrl } from './domain/useCases/GetSignedDatasetVersionDownloadUrl'
 import { SubmitGuestbookForDatafileDownload } from './domain/useCases/SubmitGuestbookForDatafileDownload'
 import { SubmitGuestbookForDatafilesDownload } from './domain/useCases/SubmitGuestbookForDatafilesDownload'
 import { SubmitGuestbookForDatasetDownload } from './domain/useCases/SubmitGuestbookForDatasetDownload'
@@ -6,6 +10,10 @@ import { SubmitGuestbookForDatasetVersionDownload } from './domain/useCases/Subm
 
 const accessRepository = new AccessRepository()
 
+const getSignedDatafileDownloadUrl = new GetSignedDatafileDownloadUrl(accessRepository)
+const getSignedDatafilesDownloadUrl = new GetSignedDatafilesDownloadUrl(accessRepository)
+const getSignedDatasetDownloadUrl = new GetSignedDatasetDownloadUrl(accessRepository)
+const getSignedDatasetVersionDownloadUrl = new GetSignedDatasetVersionDownloadUrl(accessRepository)
 const submitGuestbookForDatafileDownload = new SubmitGuestbookForDatafileDownload(accessRepository)
 const submitGuestbookForDatafilesDownload = new SubmitGuestbookForDatafilesDownload(
   accessRepository
@@ -16,6 +24,10 @@ const submitGuestbookForDatasetVersionDownload = new SubmitGuestbookForDatasetVe
 )
 
 export {
+  getSignedDatafileDownloadUrl,
+  getSignedDatafilesDownloadUrl,
+  getSignedDatasetDownloadUrl,
+  getSignedDatasetVersionDownloadUrl,
   submitGuestbookForDatafileDownload,
   submitGuestbookForDatafilesDownload,
   submitGuestbookForDatasetDownload,