+++ src/guestbooks/domain/dtos/CreateGuestbookDTO.ts
@@ -21,5 +21,8 @@ export interface CreateGuestbookDTO {
   nameRequired: boolean
   institutionRequired: boolean
   positionRequired: boolean
+  email: string
+  institution: string
+  position: string
   customQuestions: CreateGuestbookCustomQuestionDTO[]
 }