+++ test/unit/guestbooks/CreateGuestbook.test.ts
@@ -11,6 +11,9 @@ describe('CreateGuestbook', () => {
     nameRequired: true,
     institutionRequired: false,
     positionRequired: false,
+    email: 'test@gmail.com',
+    institution: 'Harvard University',
+    position: 'Researcher',
     customQuestions: [
       {
         question: "how's your day",