+++ test/unit/guestbooks/GetGuestbook.test.ts
@@ -12,6 +12,9 @@ describe('execute', () => {
     nameRequired: true,
     institutionRequired: false,
     positionRequired: false,
+    email: 'test@gmail.com',
+    institution: 'Harvard University',
+    position: 'Researcher',
     customQuestions: [],
     createTime: '2024-01-01T00:00:00Z',
     dataverseId: 34